+++ cognite/powerops/resync/config_classes/resource_collection.py
@@ -25,7 +25,6 @@ from cognite.powerops.clients.powerops_client import PowerOpsClient
 from cognite.powerops.resync.config_classes.cogshop.shop_file_config import ShopFileConfig
 from cognite.powerops.resync.config_classes.shared import ExternalId
 from cognite.powerops.resync.config_classes.to_delete import SequenceContent
-from cognite.powerops.resync.logic import clean_cdf_resources_for_diff, clean_local_resources_for_diff
 from cognite.powerops.resync.models.cdf_resources import CDFFile, CDFSequence
 from cognite.powerops.utils.cdf.calls import upsert_cognite_resources
 
@@ -365,3 +364,36 @@ def to_dm_apply(instances: list[DomainModel] | DomainModel) -> list[DomainModelA
             apply_type(**{field: value for field, value in item.dict().items() if field in apply_type.__fields__})
         )
     return apply_items
+
+
+def clean_cdf_resources_for_diff(cdf_resources: dict[ExternalId, dict]) -> None:
+    """Remove fields that are not relevant for diffing"""
+    dynamic_fields = ["last_updated_time", "created_time", "parent_id", "root_id", "data_set_id", "id"]
+    for resource in cdf_resources.values():
+        for field in dynamic_fields:
+            resource.pop(field, None)
+        # remove the metadata if it is empty
+        metadata = resource.get("metadata", {})
+        if metadata == {}:
+            resource.pop("metadata", None)
+        else:
+            for key, val in metadata.items():
+                if isinstance(val, float):
+                    # convert float to str
+                    metadata[key] = str(val)
+
+        # For sequences remove the column fields createdTime, lastUpdatedTime and metadata for each column
+        if resource.get("columns"):
+            for column in resource["columns"]:
+                column.pop("createdTime", None)
+                column.pop("lastUpdatedTime", None)
+                column.pop("metadata", None)
+
+
+def clean_local_resources_for_diff(local_resources: dict[ExternalId, dict]) -> None:
+    for resource in local_resources.values():
+        metadata = resource.get("metadata", {})
+        for key, val in metadata.items():
+            if isinstance(val, (float, int)):
+                # convert float to str
+                metadata[key] = str(val)
